make basic version of run on Windows
pty and syscall.SIGWINCH will not work on Windows. Make the command a
basic pass-through so it will compile on Windows.
