+++ cmd/binny/cli/command/run_windows.go
@@ -1,0 +1,13 @@
+package command
+
+import (
+	"os"
+	"os/exec"
+)
+
+func run(path string, args []string) error {
+	c := exec.Command(path, args...)
+	c.Stdout = os.Stdout
+	c.Stderr = os.Stderr
+	return c.Run()
+}
